refactoring initialization. better variables init. cleaning ipython notebook cells
